+++ src/commands/gistLog.ts
@@ -2,12 +2,16 @@ import * as vscode from "vscode";
 import { EXTENSION_ID } from "../constants";
 import { Gist } from "../store";
 import { newGist } from "../store/actions";
-import { GistNode } from "../tree/nodes";
+import { GistNode, GistsNode } from "../tree/nodes";
 import { fileNameToUri, openGistFile } from "../utils";
 
 const GISTLOG_URL = "https://gistlog.co";
 const GISTLOG_BLOG_FILE = "blog.md";
 const GISTLOG_CONFIG_FILE = "gistlog.yml";
+
+const createGistLogFeedUri = (login: string) => {
+  return vscode.Uri.parse(`${GISTLOG_URL}/${login}`);
+};
 
 const createGistLogUri = (gist: Gist) => {
   return vscode.Uri.parse(`${GISTLOG_URL}/${gist.owner.login}/${gist.id}`);
@@ -57,6 +61,16 @@ export async function registerGistLogCommands(
       `${EXTENSION_ID}.openGistInGistLog`,
       async (node: GistNode) => {
         const uri = createGistLogUri(node.gist);
+        vscode.env.openExternal(uri);
+      }
+    )
+  );
+
+  context.subscriptions.push(
+    vscode.commands.registerCommand(
+      `${EXTENSION_ID}.openGistLogFeed`,
+      async (node: GistsNode) => {
+        const uri = createGistLogFeedUri(node.login);
         vscode.env.openExternal(uri);
       }
     )